[#11] refactor: 데이터 constants 폴더로 이동
