+++ src/constants/data.ts
@@ -1,7 +1,45 @@
 import React from 'react';
+import mapPin from '@/assets/landing/map-pin.png';
+import location from '@/assets/landing/location.png';
+import bookmark from '@/assets/landing/bookmark.png';
+import chat from '@/assets/landing/chat.png';
+import star from '@/assets/landing/star.png';
 
 import { LuGrip, LuDog, LuWatch, LuUtensils, LuSalad, LuPencilRuler, LuLamp, LuClover, LuPill } from 'react-icons/lu';
 import { CiHome, CiShoppingBasket, CiChat1, CiBullhorn, CiUser } from 'react-icons/ci';
+
+export const LANDING_PAGE_DATA = [
+  {
+    img: mapPin,
+    title: '동네 사람들과 함께하는 구매!',
+    desc: '당신이 원하는 상품을 지금 찾아보세요.',
+    color: 'bg-primary/70',
+  },
+  {
+    img: location,
+    title: '내 주소를 기반으로 한 장소 검색',
+    desc: '동네에서 올라온 글을 확인하세요.',
+    color: 'bg-slate-400/80',
+  },
+  {
+    img: bookmark,
+    title: '고민이 되시나요?',
+    desc: '글을 북마크해서 나중에 다시 확인할 수 있어요.',
+    color: 'bg-green-600/80',
+  },
+  {
+    img: star,
+    title: '함께 한 회원들을 평가',
+    desc: '매너 평가를 통해 더 나은 커뮤니티를 만들어가요.',
+    color: 'bg-yellow-400/80',
+  },
+  {
+    img: chat,
+    title: '채팅으로 소통',
+    desc: '편하게 채팅으로 소통하세요!',
+    color: 'bg-red-300/80',
+  },
+] as const;
 
 export const NAVIGATION = [
   { path: '/home', label: '홈', icon: React.createElement(CiHome) },
@@ -28,6 +66,13 @@ export const STATUS = [
   { type: 'IN_PROGRESS', name: '모집중' },
   { type: 'COMPLETED', name: '모집완료' },
   { type: 'FAILED', name: '모집실패' },
+] as const;
+
+export const PURCHASE_STATUS_OPTIONS = [
+  { type: 'PREPARING', label: '준비중' },
+  { type: 'PROCEEDING', label: '진행중' },
+  { type: 'SUCCESS', label: '구매완료' },
+  { type: 'FAILURE', label: '구매실패' },
 ] as const;
 
 export const SORT_OPTIONS = [
