--- src/constants/options.ts
@@ -1,6 +1,0 @@
-export const PURCHASE_STATUS_OPTIONS = [
-  { type: 'PREPARING', label: '준비중' },
-  { type: 'PROCEEDING', label: '진행중' },
-  { type: 'SUCCESS', label: '구매완료' },
-  { type: 'FAILURE', label: '구매실패' },
-] as const;
